migrate references for certificate related methods to CertificateService
