+++ frontend/src/app/features/training-and-qualifications/add-edit-training/add-edit-training.component.ts
@@ -6,6 +6,7 @@ import { DATE_PARSE_FORMAT } from '@core/constants/constants';
 import { CertificateDownload, TrainingCertificate } from '@core/model/training.model';
 import { AlertService } from '@core/services/alert.service';
 import { BackLinkService } from '@core/services/backLink.service';
+import { TrainingCertificateService } from '@core/services/certificate.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
 import { TrainingCategoryService } from '@core/services/training-category.service';
 import { TrainingService } from '@core/services/training.service';
@@ -36,6 +37,7 @@ export class AddEditTrainingComponent extends AddEditTrainingDirective implement
     protected errorSummaryService: ErrorSummaryService,
     protected trainingService: TrainingService,
     protected trainingCategoryService: TrainingCategoryService,
+    protected certificateService: TrainingCertificateService,
     protected workerService: WorkerService,
     protected alertService: AlertService,
     protected http: HttpClient,
@@ -210,7 +212,7 @@ export class AddEditTrainingComponent extends AddEditTrainingDirective implement
   private uploadNewCertificate(trainingRecordResponse: any) {
     const trainingRecordId = this.trainingRecordId ?? trainingRecordResponse.uid;
 
-    return this.trainingService.addCertificateToTraining(
+    return this.certificateService.addCertificates(
       this.workplace.uid,
       this.worker.uid,
       trainingRecordId,
@@ -226,7 +228,7 @@ export class AddEditTrainingComponent extends AddEditTrainingDirective implement
             return this.formatForCertificateDownload(certificate);
           });
     this.subscriptions.add(
-      this.trainingService
+      this.certificateService
         .downloadCertificates(this.workplace.uid, this.worker.uid, this.trainingRecordId, filesToDownload)
         .subscribe(
           () => {
@@ -245,7 +247,7 @@ export class AddEditTrainingComponent extends AddEditTrainingDirective implement
 
   private deleteTrainingCertificate(files: TrainingCertificate[]) {
     this.subscriptions.add(
-      this.trainingService
+      this.certificateService
         .deleteCertificates(this.establishmentUid, this.workerId, this.trainingRecordId, files)
         .subscribe(() => {}),
     );
